refactor: ChessService에서 Dao 주입하도록 구조 변경
